pyproject.toml
[project]
name = "mini-wikipedia-rag"
version = "0.1.0"
description = "Example RAG environment using a few wikipedia articles with Llama Index"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "fastapi>=0.121.3",
    "huggingface-hub>=0.27.0",
    "jupyterlab>=4.5.6",
    "llama-index-core>=0.12.0",
    "llama-index-embeddings-azure-openai>=0.4.1",
    "llama-index-llms-azure-openai>=0.4.2",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.23.0",
    "pytest>=9.0.1",
    "requests>=2.32.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--ignore=tests/external"
# Filter out known warnings from dependencies
filterwarnings = [
    "ignore::DeprecationWarning:llama_index.*",
    "ignore:There is no current event loop:DeprecationWarning"
]
